cs_http_imt (cstasks v1)

Functions

extension_to_type(Extension, Default)

-spec extension_to_type(Extension, Default) -> Imt | Default
                           when Extension :: binary() | any(), Imt :: binary(), Default :: any().

Return the IMT for the specified file extension or Default if the file extension is unknown or invalid. Optimized for most used cases like html, png, jpg, js, json...

file_type(File, Default)

-spec file_type(File, Default) -> string() | Default when File :: iolist() | binary(), Default :: any().

Return the IMT for the specified filename. Returns Default parameter if the IMT could not be determined.

type_to_extensions(ImtType, Default)

-spec type_to_extensions(ImtType, Default) -> ImtExtensions | Default
                            when
                                ImtType :: binary() | any(),
                                ImtExtensions :: [binary()],
                                Default :: any().

Return the extension for the specified IMT or Default if the IMT is unknown or invalid.