cs_validate (cstasks v1)

Functions

safe_domain_extension(Ext)

Check domain extension Extension must contain at least one alpha numeric character. Characters are converted from upper case to lower case in the same iteration and sanitized binary is returned.

safe_http_www_path(Path)

-spec safe_http_www_path(Path) -> {ok, SanitizedPath, Extension, LastCharacter} | {error, Reason}
                            when
                                Path :: binary(),
                                SanitizedPath :: binary(),
                                Extension :: undefined | binary(),
                                LastCharacter :: undefined | integer(),
                                Reason :: atom().

Check and sanitize the specified request path. Returns the lower case path, the extension if any and the last character of the path. Those are very usefull to the caller when resolving the path. It is interesting that i couldn't find any scenario when a path using other validation rules is nedeed except in well known exploits and crappy php. With those simple rules we can quickly filter out a huge ammount of exploits scans.

safe_sub_domain_name(Hostname)

Check the specified domain name for injection So called "safe domain names" must contain only ASCII alpha-numeric characters, -, and one . before the alphanumeric extension.

A safe domain name example: sub-domain.domain-name.extension
Invalid domain names: name invalid-name -invalid-name invalid--name invalid-name. invalid..extension .invalid